fix: work with lazy.nvim config=true
